Switch to hcl4j 0.5 (based on jFlex 1.4.0)

+++ src/main/java/net/aschemann/jqassistant/plugin/hcl/api/model/HclConfiguredDescriptor.java
@@ -21,7 +21,7 @@ public interface HclConfiguredDescriptor extends HclDescriptor {
 
     void setColumn(final int column);
 
-    int getPosition();
+    long getPosition();
 
-    void setPosition(final int position);
+    void setPosition(final long position);
 }
